refactor: changed bk to tracking model
